requirements.txt
ultralytics==8.2.103
opencv-python-headless==4.10.0.84
numpy==1.26.4
scipy==1.12.0
matplotlib==3.8.4
filterpy==1.4.5
librosa==0.10.2.post1
soundfile==0.12.1
pyyaml==6.0.2
ultralytics==8.2.103
